[GUI] Synchronization estimation system
ZENZO Core v2.0 now displays syncing information in a very user-friendly manner. Now, synchronization is displayed with an estimation until completion (Based on checkpoints) along with the current block. This system collects the average syncing speed (in blocks-per-sec) of the last 10 GUI refreshes and displays the timer in the format: HH:MM:SS